Renaming and refactoring discovery test mocks
