+++ ecommerce/enterprise/tests/test_entitlements.py
@@ -10,7 +10,6 @@ from testfixtures import LogCapture
 from ecommerce.core.tests import toggle_switch
 from ecommerce.coupons.tests.mixins import CouponMixin, DiscoveryMockMixin
 from ecommerce.courses.tests.factories import CourseFactory
-from ecommerce.courses.tests.mixins import DiscoveryServiceMockMixin
 from ecommerce.enterprise.entitlements import (
     get_course_entitlements_for_learner, get_course_vouchers_for_learner, get_entitlement_voucher,
     is_course_in_enterprise_catalog
@@ -26,7 +25,7 @@ StockRecord = get_model('partner', 'StockRecord')
 
 @ddt.ddt
 @httpretty.activate
-class EntitlementsTests(EnterpriseServiceMockMixin, DiscoveryServiceMockMixin, CourseCatalogTestMixin,
+class EntitlementsTests(EnterpriseServiceMockMixin, CourseCatalogTestMixin,
                         DiscoveryMockMixin, CouponMixin, TestCase):
     def setUp(self):
         super(EntitlementsTests, self).setUp()
@@ -141,7 +140,7 @@ class EntitlementsTests(EnterpriseServiceMockMixin, DiscoveryServiceMockMixin, C
         enterprise_catalog_id = 1
         self.mock_enterprise_learner_api(entitlement_id=coupon.id)
         self.mock_enterprise_learner_entitlements_api(entitlement_id=coupon.id)
-        self.mock_discovery_api_for_catalog_contains(
+        self.mock_catalog_contains_endpoint(
             catalog_id=enterprise_catalog_id, course_run_ids=[self.course.id]
         )
 
@@ -162,7 +161,7 @@ class EntitlementsTests(EnterpriseServiceMockMixin, DiscoveryServiceMockMixin, C
             catalog_id=non_existing_coupon_id, entitlement_id=non_existing_coupon_id
         )
         self.mock_enterprise_learner_entitlements_api(entitlement_id=non_existing_coupon_id)
-        self.mock_discovery_api_for_catalog_contains(
+        self.mock_catalog_contains_endpoint(
             catalog_id=non_existing_coupon_id, course_run_ids=[self.course.id]
         )
 
@@ -194,7 +193,7 @@ class EntitlementsTests(EnterpriseServiceMockMixin, DiscoveryServiceMockMixin, C
 
         self.mock_enterprise_learner_api(entitlement_id=coupon.id)
         self.mock_enterprise_learner_entitlements_api(entitlement_id=coupon.id)
-        self.mock_discovery_api_for_catalog_contains(
+        self.mock_catalog_contains_endpoint(
             catalog_id=catalog_id, course_run_ids=[self.course.id]
         )
         self.mock_access_token_response()
@@ -242,7 +241,7 @@ class EntitlementsTests(EnterpriseServiceMockMixin, DiscoveryServiceMockMixin, C
         learner_id = 1
         self.mock_enterprise_learner_api(learner_id=learner_id)
         self.mock_learner_entitlements_api_failure(learner_id=learner_id)
-        self.mock_discovery_api_for_catalog_contains(course_run_ids=[self.course.id])
+        self.mock_catalog_contains_endpoint(course_run_ids=[self.course.id])
 
         self.mock_access_token_response()
         self._assert_get_course_entitlements_for_learner_response(
@@ -261,7 +260,7 @@ class EntitlementsTests(EnterpriseServiceMockMixin, DiscoveryServiceMockMixin, C
         learner_id = 1
         self.mock_enterprise_learner_api(learner_id=learner_id)
         self.mock_learner_entitlements_api_failure(learner_id=learner_id, status=200)
-        self.mock_discovery_api_for_catalog_contains(course_run_ids=[self.course.id])
+        self.mock_catalog_contains_endpoint(course_run_ids=[self.course.id])
 
         self.mock_access_token_response()
         self._assert_get_course_entitlements_for_learner_response(
@@ -328,7 +327,7 @@ class EntitlementsTests(EnterpriseServiceMockMixin, DiscoveryServiceMockMixin, C
         enterprise_catalog_id = 1
         self.mock_enterprise_learner_api(entitlement_id=enterprise_catalog_id)
         self.mock_enterprise_learner_entitlements_api(entitlement_id=enterprise_catalog_id)
-        self.mock_discovery_api_for_catalog_contains(
+        self.mock_catalog_contains_endpoint(
             catalog_id=enterprise_catalog_id, course_run_ids=[self.course.id]
         )
 
@@ -351,7 +350,7 @@ class EntitlementsTests(EnterpriseServiceMockMixin, DiscoveryServiceMockMixin, C
         """
         toggle_switch('use_multi_tenant_discovery_api_urls', True)
         enterprise_catalog_id = 1
-        self.mock_discovery_api_for_catalog_contains(
+        self.mock_catalog_contains_endpoint(
             catalog_id=enterprise_catalog_id, course_run_ids=[self.course.id]
         )
 
@@ -369,7 +368,7 @@ class EntitlementsTests(EnterpriseServiceMockMixin, DiscoveryServiceMockMixin, C
         the provided course is not available in the enterprise course catalog.
         """
         enterprise_catalog_id = 1
-        self.mock_discovery_api_for_catalog_contains(
+        self.mock_catalog_contains_endpoint(
             catalog_id=enterprise_catalog_id, course_run_ids=[self.course.id]
         )
 
@@ -405,7 +404,7 @@ class EntitlementsTests(EnterpriseServiceMockMixin, DiscoveryServiceMockMixin, C
         """
         toggle_switch("use_multi_tenant_discovery_api_urls", True)
         enterprise_catalog_id = 1
-        self.mock_get_catalog_contains_api_for_failure(
+        self.mock_course_runs_contains_endpoint_failure(
             course_run_ids=[self.course.id], catalog_id=enterprise_catalog_id, error=error
         )
