+++ ecommerce/extensions/api/v2/tests/views/test_coupons.py
@@ -780,7 +780,7 @@ class CouponViewSetFunctionalTest(CouponMixin, CourseCatalogTestMixin, Discovery
         })
         self.data.pop('stock_record_ids')
         course, __ = self.create_course_and_seat(course_id='dynamic/catalog/coupon')
-        self.mock_dynamic_catalog_course_runs_api(query=catalog_query, course_run=course)
+        self.mock_course_runs_endpoint(query=catalog_query, course_run=course)
 
         response = self.client.post(COUPONS_LINK, json.dumps(self.data), 'application/json')
         coupon_id = json.loads(response.content)['coupon_id']
